docs(repo): update contributing flle to mention updating changelog
ISSUES CLOSED: #21

+++ src/main/java/automenta/vivisect/timeline/BarChart.java
@@ -45,7 +45,7 @@ public class BarChart extends LineChart {
                 continue;
             }
             
-            float p = (max == min) ? 0 : (float) ((v - min) / (max - min));
+            float p = (max == min) ? 0 : (v - min) / (max - min);
             float px = width * x;
             float h = p * yScale1;
             float py = y + yScale1 - h;
